Zmiany przed zajeciami 27.06
