+++ src/main/java/syll25/tictactoe/logic/state/TxtState.java
@@ -21,7 +21,7 @@ public class TxtState implements State {
 
         try {
 
-            PrintWriter out = new PrintWriter(new FileWriter(filename));
+            PrintWriter out = new PrintWriter(new FileWriter(filename)); // to miejsce !!
             savePlayerData(player1, out);
             out.print(PLAYERS_SEPARATOR);
             savePlayerData(player2, out);
